Changes function to play round of game
Adds new function to get player choice.
This function validates that input is within the options of the game.
Game round function now returns an array containing message for round
results and a boolean true/false for win/lose.
Boolean values will be used in the playGame() function to track score.